提交，大屏增加，代码

+++ src/main/java/com/pg/screen/mapper/entity/FaultDetail.java
@@ -37,7 +37,7 @@ public class FaultDetail implements Serializable {
     /**
      * 事件序号
      */
-    private Integer eventSequenceNumber;
+    private Long eventSequenceNumber;
     /**
      * 事件时间
      */
